EOS
Tested the functionality of moving multiple motors at the same time; Using the "AccelStepper" library.
Start on implementing a function for the movement of the drill; Drilling out the charakters, following their coords.

Fix:
Fixed a missmatch of the Y and Z axis.
(Blew up a capacitor while testing, whoops)

Note:
Refactoring of the code is REQUIRED very soon, as the library offers better performance and more options.
Movement code could/will probably be rewritten completely.